Renamed base_model to base.
