+++ phy/cluster/session.py
@@ -13,7 +13,7 @@ import shutil
 
 from ..utils.logging import info, warn
 from ..utils.settings import _ensure_dir_exists
-from ..io.base_model import BaseSession
+from ..io.base import BaseSession
 from ..io.kwik.model import KwikModel
 from ..io.kwik.store_items import create_store
 from .manual.gui import ClusterManualGUI
